+++ releasenotes.md
@@ -3,6 +3,7 @@
 ## Version 0.12.0
 ### New Features
 - Address.BuildingNumber()
+- Address.Community()
 
 ## Version 0.11.0
 ### New Features